+++ crates/bid-scraper/src/lib.rs
@@ -6,6 +6,7 @@ mod slot;
 pub mod ultrasound_ws_publisher;
 
 pub mod best_bid_ws_connector;
+pub mod bid_scraper;
 pub mod bid_scraper_client;
 pub mod bid_sender;
 pub mod code_from_rbuilder;
